requirements.txt
django>=1.6,<1.9
djangorestframework<4.0
python-social-auth==0.2.9
